version 0.9.10.10 config

+++ client-module/client/src/main/java/io/nuls/client/rpc/resources/util/FileUtil.java
@@ -216,7 +216,11 @@ public final class FileUtil {
         for (File file : files) {
             if (file.isFile()) {
                 try {
-                    file.delete();
+                    boolean b = file.delete();
+                    if(!b){
+                        System.gc();
+                        file.delete();
+                    }
                 } catch (Exception e) {
                     Log.error(e);
                 }
@@ -225,7 +229,11 @@ public final class FileUtil {
             }
         }
         try {
-            folder.delete();
+            boolean b = folder.delete();
+            if(!b){
+                System.gc();
+                folder.delete();
+            }
         } catch (Exception e) {
             Log.error(e);
         }
